HCAP-1143 | Fix circular dep issue | Create new sync-user module
