+++ server/services/user.js
@@ -1,6 +1,4 @@
 const { dbClient, collections } = require('../db');
-const keycloak = require('../keycloak');
-const logger = require('../logger');
 
 const userRegionQuery = (regions, target) => {
   if (regions.length === 0) return null;
@@ -30,32 +28,9 @@ const makeUser = async ({ keycloakId, sites }) => {
   });
 };
 
-const syncUser = async ({ log }) => {
-  // Fetch all db users
-  const allUsers = await dbClient.db[collections.USERS].findDoc();
-  // Fetch all keycloak users
-  const keycloakUsers = await keycloak.getUsers(true);
-  // Sync
-  const resp = await Promise.all(
-    allUsers.map(async (user) => {
-      const keycloakUser = keycloakUsers.find((item) => item.id === user.keycloakId);
-      if (keycloakUser) {
-        const { id, ...details } = keycloakUser;
-        if (log) logger.info(`Syncing user ${user.id} and keycloak user ${id}`);
-        return dbClient.db[collections.USERS].updateDoc(user.id, {
-          userInfo: details,
-        });
-      }
-      return user;
-    })
-  );
-  return resp;
-};
-
 module.exports = {
   getUser,
   getUserSites,
   userRegionQuery,
   makeUser,
-  syncUser,
 };
